requirements.txt
absl-py==2.1.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
docker-pycreds==0.4.0
dominate==2.9.1
executing==2.0.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.2.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
sentry-sdk==2.0.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tkinterdnd2==0.3.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
visdom==0.2.4
wandb==0.16.6
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1